requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlalchemy
psycopg2-binary
redis
celery
python-jose[cryptography]
passlib[bcrypt]
python-dotenv